requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
scikit-image>=0.18.0
ffmpeg-python>=0.2.0
Flask>=2.0.0
SQLAlchemy>=1.4.0
pytest>=6.0.0
matplotlib>=3.4.0
pillow>=8.0.0
tqdm>=4.62.0
